도시에서 등록, 수정 사용자 정보 제거 및 도시 삭제 로직 수정

+++ src/main/java/com/triple/task/travel/city/application/service/CityCommandService.java
@@ -34,7 +34,12 @@ public class CityCommandService {
     }
 
     public void delete(Long cityId) {
-        // TODO 여행이 지정되지 않은 경우만 삭제
+        City city = cityQueryService.selectWithTripsBy(cityId);
+
+        if (city.hasTrip()) {
+            throw new RuntimeException();
+        }
+
         cityViewRepository.deleteCityViewsBy(cityId);
         cityRepository.deleteById(cityId);
     }
